backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pandas==2.2.2
plotly==5.22.0
scikit-learn==1.5.0
fpdf2==2.7.9
gunicorn==22.0.0
python-dotenv==1.0.1
numpy==1.26.4
